mv minimize tests into next mod::tests
